requirements.txt
redisearch==2.0.0
redis==3.5.3
slackclient==2.9.3
